Upgrade to new Lucene snapshot

+++ x-pack/plugin/old-lucene-versions/src/main/java/org/elasticsearch/xpack/lucene/bwc/codecs/index/LegacySortedSetDocValuesWrapper.java
@@ -99,7 +99,7 @@ public final class LegacySortedSetDocValuesWrapper extends SortedSetDocValues {
     }
 
     @Override
-    public long docValueCount() {
+    public int docValueCount() {
         throw new UnsupportedOperationException();
     }
